requirements.txt
# Core dependencies
flask==2.3.3
streamlit==1.28.1
pandas==2.1.1
numpy==1.24.3
requests==2.31.0

# NLP and Sentiment Analysis
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2
transformers==4.34.1
torch==2.0.1

# Data Collection
tweepy==4.14.0
praw==7.7.1
beautifulsoup4==4.12.2
selenium==4.15.0

# Financial Data
yfinance==0.2.22
alpha-vantage==2.3.1

# Visualization
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2

# Database
sqlalchemy==2.0.21
sqlite3

# Data Processing
scikit-learn==1.3.0
scipy==1.11.3

# Environment and Configuration
python-dotenv==1.0.0
configparser==6.0.0

# News APIs
newsapi-python==0.2.7

# Web scraping utilities
lxml==4.9.3
html5lib==1.1

# Development and Testing
pytest==7.4.2
jupyter==1.0.0
notebook==7.0.4

# Date and time handling
python-dateutil==2.8.2
